change bcif and cif to take from same pdbx parser
