+++ molecularnodes/io/parse/__init__.py
@@ -3,8 +3,9 @@ A subpackge which provides classes for parsing the different macromolecular data
 """
 
 from .mmtf import MMTF
-from .bcif import BCIF
-from .cif import CIF
+from .pdbx import CIF, BCIF
+# from .bcif import BCIF
+# from .cif import CIF
 from .pdb import PDB
 from .cellpack import CellPack
 from .star import StarFile
